separate user model test cases into a independent test file
